[cfe] Avoid repeated calls to .codeUnit() - use startsWith with offset instead
In AOT compiled programs we may not know whether a String is one-byte or
two-byte (internal or external) and as a result, calls to .codeUnit()
are not inlined.

The string canonicalizer uses it for two purposes: creating hashes for
substrings and comparing substrings.

=> The ladder we can replace by a call to .startsWith() with a specific
offset.

Change-Id: I31025b3be03277a9c9505ef09d9c8a41a57e7839
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/279974
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>