[analysis_server] Don't use window/showMessageRequest unless client has explicitly expressed support
LSP currently has no capability for showMessageRequest and assumed all clients support it. They do not.

Until there's a capability (and clients have been updated to pass it), we should assume there is _not_ support for this. So that it can be used in VS Code, we accept a custom flag in experimental that indicates support for this (which the current VS Code extension is passing).

When there is an official capability we should treat either as an indication of this being supported, and then further into the future we can drop the custom flag.

See https://github.com/microsoft/language-server-protocol/issues/1635.

Change-Id: I7ca36555a5a10231f754faf6ae0ba199f791be35
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/280173
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>