[dart2wasm] Omit type checks in the omit-checks benchmarking config.
The dart2wasm-omit-checks benchmarking config will pass the
`--omit-checks` option to the `compile_benchmark` script.

Detect this option and use it to enable the `--omit-type-checks`
option to dart2wasm.

As we add more options for omitting checks, we can add those to the
dart2wasm commandline we use when given the `--omit-checks` option.

Change-Id: I8aba801935888aa13357d8fd3e2f0c084c4173b5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/280091
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Aske Simon Christensen <askesc@google.com>