Flow analysis: fix handling of cast patterns.
Previously, we treated cast patterns as a simple type-check, relying
on promotion machinery to supply the correct matched value type for
the inner pattern.  In other words `P as T` was analyzed like `T() &&
P`.  This had two unforunate consequences:

(1) If the type `T` was not a subtype of the matched value type, no
    promotion occurred.  So for example, if `x` had type `int?`, then
    `if (x case var y as String?)` caused `y` to get an inferred type
    of `int?` (which is clearly not what the user wants).

(2) Any promotions triggered inside the inner pattern would propagate
    to the outer pattern.  So for example, if `x` had the type
    `Object?`, then `if (x case int _ as num)` caused `x` to be
    promoted first to `num` and then to `int`.  Although this was
    sound, it seemed to me that it was strange and unexpected
    behaviour, because the typical user expectation when performing a
    cast is that only the type being cast to should be used for
    promotion.

The solution to both of these is simple: we analyze the inner pattern
as though its matched value is distinct from the matched value
supplied to the cast, and has a static type of the cast type.  Even
though in reality we know that the two values are the same, it's sound
for flow analysis to ignore that information.

Fixing this made it possible to add tests of some other patterns flow
analysis behaviours that weren't previously testable, so I've included
several tests in this CL that aren't strictly testing the change.

Bug: https://github.com/dart-lang/sdk/issues/50419
Change-Id: I8cf864c3bf20f55406cfd74aeb8891d8e81f93e2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/280207
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>