[test_runner] Add detect-host command line flag
The --detect-host flag is provided as a convenience for running tests 
locally with named configurations. It alters the provided configurations
if needed to match the host system and architecture.

It is an error to use the flag without named configurations or as an 
option in the test_matrix.json file.

Add a the `Configuration._cloneHelper()` constructor with all parameters
as required arguments. This provides some compile time feedback when
new fields are added to the class and the clone methods haven't been 
updated.

Change-Id: I58e2ead1c3324c2cc893a4766d7daccd2787f631
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/280139
Reviewed-by: Bob Nystrom <rnystrom@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>