[cfe] Add PatternVisitor
This adds a pattern visitor and moves the matching expression into
a visitor. Also, the InferenceVisitorImpl implements the pattern
visitor and uses it for visiting patterns.

This is preparation for move the Pattern nodes to the
external AST and move pattern lowering to the constant evaluation.

Change-Id: I9a05f4c61032cfe3d823d4558922822b501c1ac1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/282105
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>