Better experiment handling in dart run
This change enables

dart run --enable-experiment :main # When using the package:binary
syntax

This is done by disabling snapshotting when experiments are enabled. We don't have a mechanism for invalidating snapshots based on the set of experiments.

And

dart run --resident main.dart # When using the resident compiler.

Here the issue was that the experiments was passed in a wrong format to the resident compiler process.

Bug: https://github.com/dart-lang/sdk/issues/51300 https://github.com/dart-lang/sdk/issues/51303
Change-Id: Id730e83bac83da9d678df36ae8ffd6e4ee44e8a8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/282281
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Sigurd Meldgaard <sigurdm@google.com>