[_fe_analyzer_shared] Add TestEnvironment for exhaustiveness testing
This replaces the StaticTypeImpl with an implementation of the shared
exhaustiveness classes for a test type system.

This will ensure that the unittests test the shared part of the code
and make it easier to handle for instance fields on Object and
non-interface types.

Change-Id: I101cdd41cf86b2847bd5959c15d4138640e34679
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/282882
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>