[dart2js] Remove exception handling from Dart2jsConstantEvaluator.
The CFE constant evaluator was rewritten to no longer throw if constant
evaluation fails, so we can get rid of the try/catch. Internally, the
CFE constant evaluator returns an AbortConstant, which is processed by
`evaluate` into an UnevaluatedConstant containing an InvalidExpression,
which is why our existing handling is sufficient.

Change-Id: I1ddc3b7dce3081b37177dd3384c5fc19c9b28be6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/283960
Reviewed-by: Nate Biggs <natebiggs@google.com>